% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConnection.R
\name{addToClassPath}
\alias{addToClassPath}
\title{Dynamically adds a path or a jar file to the classpath.}
\usage{
addToClassPath(path, packageName = NULL)
}
\arguments{
\item{path}{a character representing the path to the directory or the JAR file
if the packageName parameter is null. Otherwise, it can just be the name of the JAR file. This path
is normalized so that expressions like myJar.jar or ./extensions/myJar.jar will be processed.}

\item{packageName}{a character representing the package.}
}
\description{
This function makes it possible to add a directory or a JAR file
to the class path. If the packageName parameter is null then the urlString
parameter must be the complete path to the directory. Otherwise, it can be
the name of the JAR file and the function will find the path through the package
name. A non null packageName parameter is typically used in packages that rely
on J4R.
}
