% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rClasses.R
\name{as.float}
\alias{as.float}
\title{Cast the object into a Java float type}
\usage{
as.float(obj)
}
\arguments{
\item{obj}{a numeric or a vector of numerics}
}
\description{
Cast the object into a Java float type
}
