% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rClasses.R
\name{as.long}
\alias{as.long}
\title{Cast the object into a Java long type}
\usage{
as.long(obj)
}
\arguments{
\item{obj}{a numeric or a vector of numerics}
}
\description{
Cast the object into a Java long type
}
