% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\docType{data}
\name{bufferLength}
\alias{bufferLength}
\title{Length of the buffer when reading from the socket connection.}
\format{An object of class \code{numeric} of length 1.}
\usage{
bufferLength
}
\description{
The buffer has a length of 100Kb by default.
}
\keyword{datasets}
