% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\docType{data}
\name{cacheEnv}
\alias{cacheEnv}
\title{The cache environment of this package}
\format{An object of class \code{environment} of length 0.}
\usage{
cacheEnv
}
\description{
This environment contains the objects that enable the connection to
the gateway server.
}
\keyword{datasets}
