% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConnection.R
\name{callJavaGC}
\alias{callJavaGC}
\title{Synchronize the Java environment with the R environment}
\usage{
callJavaGC()
}
\description{
This function call the garbage collector in R and sends the list of Java references
that have been collected to the Java server. These references are then removed from
the internal map.
}
\seealso{
\href{https://sourceforge.net/p/repiceasource/wiki/J4R/}{J4R webpage}
}
