% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{checkIfClasspathContains}
\alias{checkIfClasspathContains}
\title{Check if a Library has been loaded}
\usage{
checkIfClasspathContains(myJavaLibrary)
}
\arguments{
\item{myJavaLibrary}{a character string that stands for the java library (e.g. repicea.jar)}
}
\description{
It checks if a particular library is part of the classpath.
}
