% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{checkIfExtensionsContain}
\alias{checkIfExtensionsContain}
\title{Check if a Library has been loaded}
\usage{
checkIfExtensionsContain(myJavaLibrary, packageName = NULL,
  automaticRestart = FALSE)
}
\arguments{
\item{myJavaLibrary}{a character string that stands for the java library (e.g. repicea.jar)}

\item{packageName}{a character string that stands for the package if the function is to be used
by a package (it is null by default)}

\item{automaticRestart}{a logical (FALSE by default) that enables an automatic restart of the Java
server so that the library is included in the classpath}
}
\description{
It checks if a particular library is part of the classpath.
}
