% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javaFunctions.R
\name{createArray}
\alias{createArray}
\title{Create array}
\usage{
createArray(values)
}
\arguments{
\item{class}{the class name of the instance of the array}

\item{length1}{a positive integer that defines the length of the array}

\item{length2}{an optional positive integer that defines the second dimension of the array}
}
\value{
a java.object or a java.list object if the vectorization was used.
}
\description{
Creates an array of particular class. The dimension of the
array are controled through the length1 and length2 arguments.
The argument length2 is set to 0 by default so that the function
returns a one-dimension array.
}
\details{
If any of the three parameters is a vector, then the method relies on vectorization and the
value of the function will be a java.list with all the java.object references.
}
