% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javaFunctions.R
\name{getAllValuesFromArray}
\alias{getAllValuesFromArray}
\title{Returns all the elements of a Java array}
\usage{
getAllValuesFromArray(object, affinity. = 1)
}
\arguments{
\item{object}{a java.object reference pointing to a Java array}

\item{affinity.}{an optional parameter for multithreading (see the mclapply.j4r function)}
}
\value{
either a java.list object, a vector or a matrix
}
\description{
All the elements of an array are returned. If these elements are Java instances,
then the function value is a java.list of java.object references. Otherwise,
the value is either a vector or a matrix
}
