% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javaFunctions.R
\name{getAllValuesFromListObject}
\alias{getAllValuesFromListObject}
\title{Returns all the elements of a Java instance of List}
\usage{
getAllValuesFromListObject(object, affinity. = 1)
}
\arguments{
\item{object}{a java.object that represents a List instance in Java}

\item{affinity.}{an optional parameter for multithreading (see the mclapply.j4r function)}
}
\value{
either a java.list object or an R vector
}
\description{
All the elements of a Java List instance are returned.
}
