% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javaFunctions.R
\name{getArrayLength}
\alias{getArrayLength}
\title{Return the length of an Array instance}
\usage{
getArrayLength(object, affinity. = 1)
}
\arguments{
\item{object}{a java.object instance that represents an array}

\item{affinity.}{an optional parameter for multithreading (see the mclapply.j4r function)}
}
\value{
an integer that is the length of the array
}
\description{
This method returns an integer that is the
length of the Array.
}
