% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConnection.R
\name{getClassLoaderPaths}
\alias{getClassLoaderPaths}
\title{Retrieve the paths of the current classloader}
\usage{
getClassLoaderPaths()
}
\description{
This functions returns the paths that are currently included
in the System classloader.
}
