% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConnection.R
\name{getClassLoaderURLs}
\alias{getClassLoaderURLs}
\title{Retrieve the URLs of the current classloader}
\usage{
getClassLoaderURLs()
}
\description{
This function returns the URLs that are currently included
in the System classloader.
}
\details{
This function is deprecated. Please use the getClassLoaderPaths instead.
}
