% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConfig.R
\name{getJavaArchitecture}
\alias{getJavaArchitecture}
\title{Get Java architecture}
\usage{
getJavaArchitecture()
}
\value{
the architecture, i.e. 32-Bit or 64-Bit
}
\description{
Return the architecture of the Java installation, i.e. either
32-Bit or 64-Bit. It actually returns the second slot of
the list produced by the getJavaVersion function.
}
\seealso{
getJavaVersion
}
