% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{getListOfJavaReferences}
\alias{getListOfJavaReferences}
\title{Provide a list of the Java references}
\usage{
getListOfJavaReferences()
}
\value{
a vector with the names of the objects that belong to the java.object and java.list classes.
}
\description{
The function provides the list of the Java references in the global environment.
}
