% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConnection.R
\name{getMainInstance}
\alias{getMainInstance}
\title{Return the main instance in the case of a public server}
\usage{
getMainInstance()
}
\value{
a java.object instance or null if the main instance was not set
}
\description{
An instance of a particular class can be associated to a public server. The approach
is similar to that of Py4j package, where the gateway server is just a channel to get
to a particular instance. This R function retrieves this instance.
}
