% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{getMainLoggerInstance}
\alias{getMainLoggerInstance}
\title{Provide the Reference to the Main Logger}
\usage{
getMainLoggerInstance()
}
\value{
a reference to the main Logger instance on the Java end
}
\description{
This is a Java reference for convenience.
}
