% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConfig.R
\name{getMemorySettings}
\alias{getMemorySettings}
\title{Returns the maximum, total and free memory in Mb}
\usage{
getMemorySettings()
}
\value{
a data.frame object with the maximum, total and free memory in Mb.
}
\description{
This function calls the Runtime static methods maxMemory(),
totalMemory() and freeMemory(). The results are divided by
1024 in order to report the memory sizes in Mb.
}
