% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConnection.R
\name{getNbConnections}
\alias{getNbConnections}
\title{The number of connections to the server}
\usage{
getNbConnections()
}
\value{
the number of sockets connected to the server
}
\description{
The number of connections to the server
}
