% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConnection.R
\name{getNbInstancesInInternalMap}
\alias{getNbInstancesInInternalMap}
\title{Return the number of instances stored in the
internal map of the Java server}
\usage{
getNbInstancesInInternalMap()
}
\value{
an integer
}
\description{
Return the number of instances stored in the
internal map of the Java server
}
