% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javaFunctions.R
\name{getValueFromArray}
\alias{getValueFromArray}
\title{Get a value from an array}
\usage{
getValueFromArray(object, ..., affinity. = 1)
}
\arguments{
\item{object}{a java.object that represents an array}

\item{...}{a series of integers that correspond to the index of the value. Note that in Java
the first index is 0}

\item{affinity.}{an optional parameter for multithreading (see the mclapply.j4r function)}
}
\value{
the value at the location
}
\description{
This function returns the value at location given by
the index parameter.
}
