% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javaFunctions.R
\name{isArray}
\alias{isArray}
\title{Check if the java.object instance represents an Array}
\usage{
isArray(object)
}
\arguments{
\item{object}{a java.object instance}
}
\value{
a logical
}
\description{
This function returns true if the Java instance represented
by this java.object is an Array.
}
