% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConnection.R
\name{isConnectedToJava}
\alias{isConnectedToJava}
\title{Checks if the Java server is running}
\usage{
isConnectedToJava()
}
\value{
a logical
}
\description{
This is done by checking f the socket connection to the JVM exists.
}
