% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javaFunctions.R
\name{isJavaArray}
\alias{isJavaArray}
\title{Check if the java.object instance represents an Array}
\usage{
isJavaArray(object)
}
\arguments{
\item{object}{a java.object instance}
}
\description{
This function returns true if the Java instance represented
by this java.object is an Array.
}
\details{
This function is deprecated. Please use the is.JavaArray instead.
}
