% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConfig.R
\name{j4r.config.registerEnvironment}
\alias{j4r.config.registerEnvironment}
\title{Register an environment to J4R}
\usage{
j4r.config.registerEnvironment(envir)
}
\arguments{
\item{envir}{an environment. Has no effect if the argument is null, not an environment or is already
included in the list of environments in J4R.}
}
\value{
a logical, true if the environment has been registered or false otherwise
}
\description{
Register an additional environment that will be scanned for Java references when
calling the callJavaGC.
}
