% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConfig.R
\name{j4r.config.removeEnvironment}
\alias{j4r.config.removeEnvironment}
\title{Remove an environment to J4R}
\usage{
j4r.config.removeEnvironment(envir)
}
\arguments{
\item{envir}{an environment. Has no effect if the argument is null, not an environment or is not
included in the list of environments in J4R.}
}
\value{
a logical, true if the environment has been removed or false in all other cases
}
\description{
Remove an environment that will be scanned for Java references when
calling the callJavaGC.
}
