% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConfig.R
\name{j4r.config.setDefaultJVMMemorySize}
\alias{j4r.config.setDefaultJVMMemorySize}
\title{Set a default memory size for the Java Virtual Machine}
\usage{
j4r.config.setDefaultJVMMemorySize(defaultJVMMemory)
}
\arguments{
\item{defaultJVMMemory}{the number of Mb for the JVM (must be equal to or greater than 50).
If set to NULL, this option has no effect.}
}
\description{
Allows to specify a default JVM size in Mb so that the option
memorySize in hte connectToJava function does not need to be
used.
}
