% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConfig.R
\name{j4r.config.setVerbose}
\alias{j4r.config.setVerbose}
\title{Enabling/disabling Verbose}
\usage{
j4r.config.setVerbose(verbose)
}
\arguments{
\item{verbose}{a logical}
}
\description{
It enables or diasble the verbose in hte J4R package.
By default, the verbose is disabled.
}
