% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConnection.R
\docType{data}
\name{J4R_Server_Version}
\alias{J4R_Server_Version}
\title{The current version of the J4R Java server}
\format{
An object of class \code{character} of length 1.
}
\usage{
J4R_Server_Version
}
\description{
The current version of the J4R Java server
}
\keyword{datasets}
