% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{killJava}
\alias{killJava}
\title{Kills the JVM and removes all the java.object instances from the environment}
\usage{
killJava()
}
\description{
This function should be used when the control over the Java server has been lost.
For instance, if the server is executing an endless task, one might use the
cancelJavaTask function. However, if it does not work, then the JVM can be killed
using this function. All references to Java objects are then lost.
}
