% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rClasses.R
\name{length.java.list}
\alias{length.java.list}
\title{Override the default length function}
\usage{
\method{length}{java.list}(x)
}
\arguments{
\item{x}{a java.list instance}
}
\value{
the length of the inner list
}
\description{
A java.list class is an environment containing an inner list. The
length of this inner list is returned by this function.
}
