% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rClasses.R
\name{length.java.object}
\alias{length.java.object}
\title{Override the default length function}
\usage{
\method{length}{java.object}(x)
}
\arguments{
\item{x}{a java.object instance}
}
\value{
1
}
\description{
A java.object class is a list by definition. However, its length
is 1.
}
