% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\docType{data}
\name{maxVectorLength}
\alias{maxVectorLength}
\title{Maximum length of the vector in the parameters.}
\format{An object of class \code{numeric} of length 1.}
\usage{
maxVectorLength
}
\description{
A maximum length of the vector is set in order to avoid buffer size issues when reading
}
\keyword{datasets}
