% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rClasses.R
\name{print.java.list}
\alias{print.java.list}
\title{Print a java.list object}
\usage{
\method{print}{java.list}(x, ...)
}
\arguments{
\item{x}{a java.list instance}

\item{...}{additional parameters for consistent overriding}
}
\description{
The java.object instances that are included in
this list are displayed up to a maximum number.
}
