% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4jFunctions.R
\name{setJavaExtensionPath}
\alias{setJavaExtensionPath}
\title{This function sets a path for eventual extensions, i.e. jar files.}
\usage{
setJavaExtensionPath(path)
}
\arguments{
\item{path}{the path to the jar files to be loaded by the Java classloader}
}
\value{
nothing
}
\description{
This function sets a path for eventual extensions, i.e. jar files.
}
\examples{
setJavaExtensionPath("/home/fortin/myExternalLibraries")

}
