% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{setJavaField}
\alias{setJavaField}
\title{Set the value of a public field}
\usage{
setJavaField(source, fieldName, value)
}
\arguments{
\item{source}{this should be either a java.list instance or a single java.object instance for non-static methods or
a string representing the Java class name in case of static method}

\item{fieldName}{the name of the field to be set}

\item{value}{the new value of the field}
}
\description{
This function sets a particular field, which can be either static or not. If the field is static,
the source should be a valid class name.
}
