% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConfig.R
\name{setJavaPath}
\alias{setJavaPath}
\title{Set the path to Java}
\usage{
setJavaPath(path)
}
\arguments{
\item{path}{the complete path to Java as in the example below}
}
\description{
This is an option function that makes it possible to set the
JAVA environmental variable in R, if it is not already set.
It first tests if the path ends with java and if it is actually
a file.
}
\examples{
# setJavaPath("/usr/lib/jvm/java-8-openjdk-amd64/bin/java")  ### not run

}
