% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{setLogLevel}
\alias{setLogLevel}
\title{Set the Logging Level}
\usage{
setLogLevel(level)
}
\arguments{
\item{level}{a String among these: SEVERE, WARNING, INFO,
FINE, FINER, and FINEST (case sensitive)}
}
\value{
nothing
}
\description{
The level refers to the Java nomenclature: SEVERE, WARNING, INFO,
FINE, FINER, and FINEST.
}
