% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javaFunctions.R
\name{setValueInArray}
\alias{setValueInArray}
\title{Set a value in an array}
\usage{
setValueInArray(object, value, index = NULL, affinity. = 1)
}
\arguments{
\item{object}{a java.object that represents an array}

\item{value}{the value to be set}

\item{index}{the index of the location at which the value is set. Note that in Java
the first index is 0. If this argument is set to NULL, then it is assumed that the value
is set to index 0. In case of vectorization, the values are set from 0 to length(value) - 1
if this argument is left to NULL.}

\item{affinity.}{an optional parameter for multithreading (see the mclapply.j4r function)}
}
\description{
This function sets the value at the location given by the index
parameter. It relies on the reflexive methods the Java class Array.
}
