% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConnection.R
\name{shutdownClient}
\alias{shutdownClient}
\title{Shut down R client}
\usage{
shutdownClient()
}
\description{
This function shuts down the client. If the server is private, it is also shut down.
}
\seealso{
\href{https://sourceforge.net/p/repiceasource/wiki/J4R/}{J4R webpage}
}
