% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{shutdownJava}
\alias{shutdownJava}
\title{Shut down Java}
\usage{
shutdownJava()
}
\description{
This function shuts down Java and the gateway server.
}
\seealso{
\href{https://sourceforge.net/p/repiceasource/wiki/J4R/}{J4R webpage}
}
