% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rClasses.R
\name{[.float}
\alias{[.float}
\title{Override the subsetting for float instance}
\usage{
\method{[}{float}(x, y)
}
\arguments{
\item{x}{a float instance}

\item{y}{a numeric that stands for the index}
}
\value{
a float instance that is a subset of x
}
\description{
This function is required. Otherwise the float class
is lost after the subsetting.
}
