% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rClasses.R
\name{[.java.list}
\alias{[.java.list}
\title{Create a subsetting operator for the java.list class.}
\usage{
\method{[}{java.list}(x, y)
}
\arguments{
\item{x}{a java.list instance}

\item{y}{either a numeric or a character}
}
\value{
a java.list that contains a subset of the original java.list
}
\description{
The subsetting operator is delegated to the inner list.
}
