% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rClasses.R
\name{[.long}
\alias{[.long}
\title{Override the subsetting for long instances}
\usage{
\method{[}{long}(x, y)
}
\arguments{
\item{x}{a long instance}

\item{y}{a numeric that stands for the index}
}
\value{
a long instance that is a subset of x
}
\description{
This function is required. Otherwise the long class
is lost after the subsetting.
}
