% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{addUrlToClassPath}
\alias{addUrlToClassPath}
\title{Dynamically adds an url to the classpath.}
\usage{
addUrlToClassPath(urlString)
}
\description{
This function makes it possible to add a directory or a JAR file
to the class path.
}
\details{
@param urlString a character representing the complete path to the directory or the JAR file
}
