% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javaFunctions.R
\name{as.JavaArray}
\alias{as.JavaArray}
\title{Create a Java array from an R array}
\usage{
as.JavaArray(values)
}
\arguments{
\item{values}{a vector or a matrix}
}
\value{
a java.object that points a Java array
}
\description{
Converts an R array into a Java array.
}
