% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4jFunctions.R
\name{callJavaGC}
\alias{callJavaGC}
\title{This function synchronizes the Java environment with the R environment. Objects that
are removed from the R environment are not automatically removed from the Java
environment. This function scans the R environment for the java.object instance and
commands the gateway server to get rid of the Java instances that are not longer referred
to in the R environment.}
\usage{
callJavaGC()
}
\description{
To avoid a memory leak, the function should be called on a regular basis.
}
