% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4jFunctions.R
\name{callJavaMethod}
\alias{callJavaMethod}
\title{This method calls a public method in a particular class of object. If the javaObject parameters or the additional
parameters (...) include vectors, the method is called several times and a vector of primitive or a list of java
instances can be returned.}
\usage{
callJavaMethod(javaObject, methodName, ...)
}
\arguments{
\item{javaObject}{this should be either a list of instances or a single instance of java.object.}

\item{methodName}{the name of the method}

\item{...}{the parameters of the method}
}
\value{
It depends on the method. It can return a primitive type (or a vector of primitive), a Java instance (or a list of Java instances) or nothing at all.
}
\description{
This method calls a public method in a particular class of object. If the javaObject parameters or the additional
parameters (...) include vectors, the method is called several times and a vector of primitive or a list of java
instances can be returned.
}
\examples{
### starting Java
connectToJava()

### creating an empty ArrayList object
myList <- createJavaObject("java.util.ArrayList")

### adding 3 to the list
callJavaMethod(myList, "add", 3)

### shutting down Java
shutdownJava()

}
