% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4jFunctions.R
\name{connectToJava}
\alias{connectToJava}
\title{This function connects the R environment to a gateway server that runs in Java.
The extension path must be set before calling this function. See the setJavaExtensionPath
function.}
\usage{
connectToJava(port = 18011, local = TRUE)
}
\arguments{
\item{port}{the local port (the port is set to 18011 by default)}

\item{local}{for debugging only}
}
\value{
nothing
}
\description{
This function connects the R environment to a gateway server that runs in Java.
The extension path must be set before calling this function. See the setJavaExtensionPath
function.
}
