% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{connectToJava}
\alias{connectToJava}
\title{Connect to Java environment}
\usage{
connectToJava(
  port = 18011,
  extensionPath = NULL,
  memorySize = NULL,
  debug = FALSE
)
}
\arguments{
\item{port}{the local port (the port is set to 18011 by default)}

\item{extensionPath}{the path to jar files that can be loaded by the system classloader}

\item{memorySize}{the memory size of the Java Virtual Machine in Mb (if not specified, the JVM runs with the default memory size)}

\item{debug}{for debugging only (should be left as is)}
}
\value{
nothing
}
\description{
This function connects the R environment to a gateway server that runs in Java.
}
