% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{connectToJava}
\alias{connectToJava}
\title{Connect to Java environment}
\usage{
connectToJava(port = 18011, local = TRUE)
}
\arguments{
\item{port}{the local port (the port is set to 18011 by default)}

\item{local}{for debugging only (should be left as is)}
}
\value{
nothing
}
\description{
This function connects the R environment to a gateway server that runs in Java.
The extension path must be set before calling this function. See setJavaExtensionPath.
}
