% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{createJavaObject}
\alias{createJavaObject}
\title{Create Java objects}
\usage{
createJavaObject(class, ...)
}
\arguments{
\item{class}{the Java class of the object (e.g. java.util.ArrayList)}

\item{...}{the parameters to be passed to the constructor of the object}
}
\value{
a java.object or java.list instance in the R environment
}
\description{
This function creates one or many object of a particular class. If the parameters
contain vectors, then a series of instances of this class can be created.
}
\examples{
### starting Java
connectToJava()

### creating an empty ArrayList object
createJavaObject("java.util.ArrayList")

### creating an ArrayList instance with initial capacity of 3
createJavaObject("java.util.ArrayList", as.integer(3))

### creating two ArrayList with different capacities
createJavaObject("java.util.ArrayList", c(as.integer(3), as.integer(4)))

### shutting down Java
shutdownJava()

}
\seealso{
\href{https://sourceforge.net/p/repiceasource/wiki/J4R/}{J4R webpage}
}
