% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javaFunctions.R
\name{getAllValuesFromArray}
\alias{getAllValuesFromArray}
\title{Returns all the elements of an array}
\usage{
getAllValuesFromArray(object)
}
\arguments{
\item{object}{a java.object that represents a Java List instance}
}
\value{
either a java.arraylist object or an R vector
}
\description{
All the elements of an array are returned.
}
