% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javaFunctions.R
\name{getAllValuesFromListObject}
\alias{getAllValuesFromListObject}
\title{Returns all the elements of a Java instance of List}
\usage{
getAllValuesFromListObject(object)
}
\arguments{
\item{object}{a java.object that represents a Java List instance}
}
\value{
either a java.list object or an R vector
}
\description{
All the elements of a Java List instance are returned.
}
