% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javaFunctions.R
\name{getArrayLength}
\alias{getArrayLength}
\title{Return the length of an Array instance}
\usage{
getArrayLength(object)
}
\arguments{
\item{object}{a java.object instance that represents an array}
}
\value{
an integer that is the length of the array
}
\description{
This method returns an integer that is the
length of the Array.
}
