% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{getJavaField}
\alias{getJavaField}
\title{Get the value of a public field}
\usage{
getJavaField(source, fieldName, affinity = 1)
}
\arguments{
\item{source}{this should be either a java.list instance or a single java.object instance for non-static methods or
a string representing the Java class name in case of static method}

\item{fieldName}{the name of the field to be set}

\item{affinity}{a parameter used by the mclapply.j4r function in case of multithreading.}
}
\description{
This function gets the value of a particular field, which can be either static or not. If the field is static,
the source should be a valid class name.
}
\details{
When the source is a java.object instance, this function can be substituted for the $ operator.
}
