% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{getListOfJavaReferences}
\alias{getListOfJavaReferences}
\title{Provide a list of the Java references}
\usage{
getListOfJavaReferences(envir = .GlobalEnv)
}
\arguments{
\item{envir}{the environment to be scanned for java.object and java.list instances. By default, it is the global
environment}
}
\value{
a vector with the names of the instances
}
\description{
The function provides the list of the Java references in an environment.
}
\details{
By default this function provides the Java reference in the current environment. If
there is no Java references then the value of the function is an empty list. If
just.names is set to true, the value is a vector with the names of the instances. If false,
then the function returns a list with the instances.
}
