% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javaFunctions.R
\name{getValueFromArray}
\alias{getValueFromArray}
\title{Get a value from an array}
\usage{
getValueFromArray(object, index)
}
\arguments{
\item{object}{a java.object that represents an array}

\item{index}{the index of the location at which the value is retrieved. Note that in Java
the first index is 0}
}
\value{
the value at the location
}
\description{
This function returns the value at location given by
the index parameter.
}
