% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConnection.R
\name{killJava}
\alias{killJava}
\title{Force the JVM to shut down}
\usage{
killJava()
}
\description{
This is the not so gentle way to exit the JVM.
}
\details{
In case the JVM is stuck and does not respond to interrupt. It is possible
to force the shutdown through this function.
}
