% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{mclapply.j4r}
\alias{mclapply.j4r}
\title{Using multithreading with J4R}
\usage{
mclapply.j4r(X, FUN)
}
\arguments{
\item{X}{a vector or a list as per the mclapply function}

\item{FUN}{a two-argument function. The first argument is called by
the mclapply function and the second argument must be used in the calls
to the createJavaObject, callJavaMethod, getJavaField and setJavaField
function.}
}
\description{
Applies the mclapply function in the context of
the J4R package.
}
\details{
Multithreading a function requires that the Java code is
thread safe. The server must listen to  at least two ports.
If it is started with a single port, this function will reduce
to a single thread application. Each port is assigned to a thread.
}
