% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{mclapply.j4r}
\alias{mclapply.j4r}
\title{Using multithreading with J4R}
\usage{
mclapply.j4r(X, FUN)
}
\arguments{
\item{X}{a vector of numerics}

\item{FUN}{a two-argument function. The first argument is called by
the mclapply function and the second argument defines the affinity and MUST
be used in all the calls to the createJavaObject, callJavaMethod,
getJavaField and setJavaField functions.}
}
\description{
Applies the mclapply function in the context of
the J4R package.
}
\details{
Multithreading a function requires that the Java code is
thread safe. The server must listen to at least two ports.
Otherwise, this function will reduce to a single thread.
Each port is given an affinity to a thread, so that each Java thread
matches an R thread. NOTE: the multithreading is not
available on Windows. In such a case, the function will proceed
in a single thread.
}
\seealso{
mclapply in the parallel package
}
