% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rFunctions.R
\name{setJavaExtensionPath}
\alias{setJavaExtensionPath}
\title{Set a path for jar extensions.}
\usage{
setJavaExtensionPath(path)
}
\arguments{
\item{path}{the path to the jar files to be loaded by the Java classloader}
}
\description{
This function sets a path for eventual extensions, i.e. jar files. These
extensions are loaded through a custom classloader.
}
\examples{
setJavaExtensionPath("/home/fortin/myExternalLibraries")
connectToJava()
## your code ##
shutdownJava()

}
\seealso{
\href{https://sourceforge.net/p/repiceasource/wiki/J4R/}{J4R webpage}
}
