% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j4rConfig.R
\name{setJavaPath}
\alias{setJavaPath}
\title{Set the path to Java}
\usage{
setJavaPath(path)
}
\arguments{
\item{path}{the complete path to Java as in the example below. The file.path function should be used to
define the path}
}
\description{
This is an option function that makes it possible to set the
JAVA environment variable in R, if it is not already set.
It first tests if the path ends with java or java.exe and
if it is actually a file. Note that if an empty character is
passed to this function. It resets the JAVA environment variable.
}
\examples{
myPath <- file.path("C:","Program Files (x86)","Java", "jre1.8.0_221", "bin", "java.exe")
# setJavaPath(myPath)  ### not run

}
\seealso{
file.path
}
