% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/javaFunctions.R
\name{setValueInArray}
\alias{setValueInArray}
\title{Set a value in an array}
\usage{
setValueInArray(object, value, index)
}
\arguments{
\item{object}{a java.object that represents an array}

\item{value}{the value to be set}

\item{index}{the index of the location at which the value is retrieved. Note that in Java
the first index is 0}
}
\description{
This function sets the value at the location given by the index
parameter. It relies on the reflexive methods the Java class Array.
}
